Refactor: follow 기능 패키지 구조 변경
-분산된 follow 기능을 follow 패키지로 이동
