+++ src/main/java/com/superposition/follow/domain/entity/ArtistFollow.java
@@ -1,4 +1,4 @@
-package com.superposition.artist.domain.entity;
+package com.superposition.follow.domain.entity;
 
 import lombok.Getter;
