+++ src/main/java/com/superposition/follow/domain/mapper/ArtistFollowMapper.java
@@ -1,12 +1,8 @@
-package com.superposition.artist.domain.mapper;
+package com.superposition.follow.domain.mapper;
 
-import com.superposition.artist.domain.entity.ArtistFollow;
-import com.superposition.artist.dto.ArtistFollowDto;
-import com.superposition.artist.dto.ArtistInfo;
-import com.superposition.artist.dto.ResponseArtist;
-import com.superposition.artist.dto.ResponseDisplayArtist;
+import com.superposition.follow.domain.entity.ArtistFollow;
+import com.superposition.follow.dto.ArtistFollowDto;
 import java.util.List;
-import java.util.Optional;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
