+++ src/main/java/com/superposition/user/controller/UserController.java
@@ -1,11 +1,8 @@
 package com.superposition.user.controller;
 
-import com.superposition.artist.dto.ArtistInfo;
-import com.superposition.artist.service.ArtistFollowService;
 import com.superposition.user.dto.LoginResponse;
 import com.superposition.user.dto.RequestUserInfo;
 import com.superposition.user.service.UserService;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -21,7 +18,6 @@ import javax.validation.Valid;
 @RequiredArgsConstructor
 public class UserController {
     private final UserService userService;
-    private final ArtistFollowService artistFollowService;
 
     @PostMapping(value = "/signup")
     public ResponseEntity<LoginResponse> signup(@RequestBody @Valid RequestUserInfo userInfo){
@@ -59,11 +55,5 @@ public class UserController {
     public ResponseEntity<?> regenerateToken(@AuthenticationPrincipal UserDetails accessToken) {
         return userService.regenerateToken(
                 SecurityContextHolder.getContext().getAuthentication().getName());
-    }
-
-    @GetMapping(value = "/artist/follow")
-    @ResponseStatus(HttpStatus.OK)
-    public List<? extends ArtistInfo> getFollowArtists(@AuthenticationPrincipal UserDetails user) {
-        return artistFollowService.getFollowArtistsBy(user.getUsername());
     }
 }
